rename `iframe.ts` -> `embeddable.ts`
